fix undef behavior

+++ voxel_engine/src/utils.cpp
@@ -2,7 +2,7 @@
 
 size_t hash_ivec3(ivec3 const& key) {
     // Using random primes
-    return (((((456818903 + key.x) * 832251403) + key.y) * 1349392157) + key.z) * 1866190769;
+    return (((((456818903U + (uint)key.x) * 832251403U) + (uint)key.y) * 1349392157U) + (uint)key.z) * 1866190769U;
 }
 
 size_t hash_ivec3(ivec3 const& key, int nonce) {
@@ -11,7 +11,7 @@ size_t hash_ivec3(ivec3 const& key, int nonce) {
 
 size_t hash_ivec4(ivec4 const& key) {
     // Using random primes
-    return ((((((456818903 + key.x) * 832251403) + key.y) * 1349392157) + key.z) * 1866190769 + key.w)*74709703;
+    return ((((((456818903U + (uint)key.x) * 832251403U) + (uint)key.y) * 1349392157U) + (uint)key.z) * 1866190769U + (uint)key.w)*74709703U;
 }
 
 void write_integer(byte* buffer, unsigned index, int integer) {
